Move transformers
Down into the prospective submodule. It's not important enough or independent enough to be in the top-level module. It generates prospective recipes, so let's move it here.
